+++ lib/rdstation-ruby-client.rb
@@ -15,6 +15,8 @@ require 'rdstation/fields'
 require 'rdstation/webhooks'
 require 'rdstation/segmentations'
 require 'rdstation/emails'
+require 'rdstation/analytics'
+require 'rdstation/landing_pages'
 
 # Error handling
 require 'rdstation/error'